+++ .github/workflows/ci.yaml
@@ -103,18 +103,21 @@ jobs:
           sed "s|namespace: \"airflow-dev\"|namespace: \"$NAMESPACE\"|g" \
             helm/kueue-resources.yaml | kubectl apply -f -
 
-      - name: Run pytest
+      - name: Run unit tests
         run: |
-          uv run pytest tests/ \
+          uv run pytest tests/unit -v
+
+      - name: Run integration tests
+        run: |
+          uv run pytest tests/integration \
             --namespace $NAMESPACE \
             --chart-name $CHART_NAME \
             --expect-custom-image \
-            -k "not test_logs_pvc_is_bound and not test_e2e_hello_world and not test_kueue_e2e" \
             -v
 
       - name: Run E2E tests
         run: |
-          uv run pytest tests/test_deployment.py::test_kueue_e2e_gates_second_pod \
+          uv run pytest tests/e2e/test_kueue_gating.py \
             --namespace $NAMESPACE \
             --chart-name $CHART_NAME \
             --kueue-namespace kueue-system \
